Refactor: use luarocks for lua-zip and lua-yajl
Using the former one should mean we no longer need to compile it into our
MacOs application, and as it is a different one (Brimworks rather
than Kepler project, as is the lua-yajl module) that does not use the
zziplib/libzzip library it should save us some hassle as that library is
causing problems with the MacOS CI at the moment.

As we are *adding* the modules directly into the same directory as the
Mudlet executable it does not matter if the current CI builds of that
do not call upon it yet. I.e. this PR can go into the installers repo
before we modify the main repo to use it...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>